% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_tidygraph}
\alias{as_tidygraph}
\alias{as_tidygraph.cooccurrence}
\title{Convert to tidygraph}
\usage{
as_tidygraph(x, ...)

\method{as_tidygraph}{cooccurrence}(x, ...)
}
\arguments{
\item{x}{A \code{cooccurrence} data frame.}

\item{...}{Ignored.}
}
\value{
A \code{tbl_graph} object.
}
\description{
Creates a \code{tbl_graph} from a \code{cooccurrence} edge list.
}
