% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movies}
\alias{movies}
\title{IMDB movie metadata (1970-2024)}
\format{
A data frame with 1,000 rows and 7 variables:
\describe{
\item{tconst}{IMDB title identifier (e.g. \code{"tt0068646"}).}
\item{primaryTitle}{Movie title.}
\item{startYear}{Release year (integer).}
\item{genres}{Comma-separated genre labels (e.g. \code{"Crime,Drama"}).}
\item{decade}{Release decade as a character string (e.g. \code{"1970s"}).}
\item{averageRating}{IMDB average user rating.}
\item{numVotes}{Number of IMDB user votes.}
}
}
\source{
\url{https://developer.imdb.com/non-commercial-datasets/}
}
\usage{
movies
}
\description{
A sample of 1,000 highly-rated IMDB movies (rating >= 7.0, >= 1,000 votes)
released between 1970 and 2024. The \code{genres} column is comma-delimited
and suitable for use as the \code{field} argument to \code{\link{cooccurrence}}.
}
\examples{
head(movies)
cooccurrence(movies, field = "genres", sep = ",", similarity = "jaccard")
}
\keyword{datasets}
