% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cooccurrence}
\alias{print.cooccurrence}
\title{Print a cooccurrence edge list}
\usage{
\method{print}{cooccurrence}(x, n = 10L, ...)
}
\arguments{
\item{x}{A \code{cooccurrence} data frame.}

\item{n}{Integer. Number of rows to show. Default 10.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a cooccurrence edge list
}
