% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cooccurrence}
\alias{summary.cooccurrence}
\title{Summarise a cooccurrence network}
\usage{
\method{summary}{cooccurrence}(object, ...)
}
\arguments{
\item{object}{A \code{cooccurrence} data frame.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summarise a cooccurrence network
}
